user, course

+++ src/modules/course/presentation/controller.ts
@@ -1,3 +1,4 @@
+import { ControllerBase } from "@core/presentation/controller-base";
 import {
   CourseGetAll,
   CourseGetById,
@@ -5,35 +6,32 @@ import {
   CourseSave,
 } from "@course/application";
 import { Course, CourseProperties } from "@course/domain/course";
-import { validate, ValidationError } from "class-validator";
 import { Request, Response } from "express";
 import { v4 as uuidv4 } from "uuid";
 
-import { CourseRequestDto } from "./dtos/course-request.dto";
+import { CourseCreateDto } from "./dtos/course-create.dto";
+import { CourseDeleteDto } from "./dtos/course-delete.dto";
+import { CourseGetByIdDto } from "./dtos/course-get-by-id";
+import { CourseGetByPageDto } from "./dtos/course-get-by-page";
 import { CourseResponseDto } from "./dtos/course-response.dto";
+import { CourseUpdateDto } from "./dtos/course-update.dto";
 
-export class CourseController {
+export class CourseController extends ControllerBase {
   constructor(
     private readonly courseSave: CourseSave,
     private readonly courseGetAll: CourseGetAll,
     private readonly courseGetById: CourseGetById,
     private readonly courseGetByPage: CourseGetByPage
-  ) {}
+  ) {
+    super();
+  }
 
   async create(req: Request, res: Response) {
     const { title, slug } = req.body;
 
-    const dtoRequest = new CourseRequestDto();
-    dtoRequest.title = title;
-    dtoRequest.slug = slug;
-
-    const errors: ValidationError[] = await validate(dtoRequest, {
-      whitelist: true,
-      forbidNonWhitelisted: true,
-      forbidUnknownValues: true,
-    });
-    if (errors.length > 0) {
-      return res.status(411).json({ errors });
+    const errors = await this.validateParameters(CourseCreateDto, req.body);
+    if (errors) {
+      return res.status(400).json(errors);
     }
 
     const courseId = uuidv4();
@@ -53,6 +51,14 @@ export class CourseController {
     const { courseId } = req.params;
     const body = req.body;
 
+    const errors = await this.validateParameters(CourseUpdateDto, {
+      ...req.body,
+      courseId,
+    });
+    if (errors) {
+      return res.status(400).json(errors);
+    }
+
     const course = await this.courseGetById.execute(courseId);
     course.update(body);
     const valueReturned = await this.courseSave.execute(course);
@@ -61,6 +67,13 @@ export class CourseController {
 
   async delete(req: Request, res: Response) {
     const { courseId } = req.params;
+
+    const errors = await this.validateParameters(CourseDeleteDto, {
+      courseId,
+    });
+    if (errors) {
+      return res.status(400).json(errors);
+    }
 
     const course = await this.courseGetById.execute(courseId);
     course.delete();
@@ -75,17 +88,33 @@ export class CourseController {
 
   async getById(req: Request, res: Response) {
     const { courseId } = req.params;
+
+    const errors = await this.validateParameters(CourseGetByIdDto, {
+      courseId,
+    });
+    if (errors) {
+      return res.status(400).json(errors);
+    }
+
     const course = await this.courseGetById.execute(courseId);
     res.json(CourseResponseDto.fromDomainToResponse(course));
   }
 
   async getByPage(req: Request, res: Response) {
     const { page, pageSize } = req.params;
+
+    const errors = await this.validateParameters(CourseGetByPageDto, {
+      page: Number(page),
+      pageSize: Number(pageSize),
+    });
+    if (errors) {
+      return res.status(400).json(errors);
+    }
+
     const courses = await this.courseGetByPage.execute(
       parseInt(page),
       parseInt(pageSize)
     );
     res.json(CourseResponseDto.fromDomainToResponse(courses));
-    //return await this.courseGetByPage.execute(page, pageSize);
   }
 }
